pyproject.toml
[project]
name = "notificate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "aiogram>=3.21.0",
    "aio-pika>=9.4.0",
    "fastapi>=0.115.14",
    "jinja2>=3.1.6",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.29.0",
]
